Move ItemFactory and ItemConfigs to a new package
