+++ even-more-fish-plugin/src/main/java/com/oheers/fish/baits/Bait.java
@@ -13,8 +13,8 @@ import com.oheers.fish.messages.ConfigMessage;
 import com.oheers.fish.messages.EMFListMessage;
 import com.oheers.fish.messages.EMFSingleMessage;
 import com.oheers.fish.messages.abstracted.EMFMessage;
-import com.oheers.fish.utils.ItemFactory;
-import com.oheers.fish.utils.items.DisplayNameItemConfig;
+import com.oheers.fish.items.ItemFactory;
+import com.oheers.fish.items.configs.DisplayNameItemConfig;
 import dev.dejvokep.boostedyaml.block.implementation.Section;
 import net.kyori.adventure.text.Component;
 import org.bukkit.Location;
