+++ even-more-fish-plugin/src/main/java/com/oheers/fish/items/configs/PotionMetaItemConfig.java
@@ -1,4 +1,4 @@
-package com.oheers.fish.utils.items;
+package com.oheers.fish.items.configs;
 
 import com.oheers.fish.EvenMoreFish;
 import com.oheers.fish.FishUtils;
